+++ packages/grid/src/features/standard-table/stories/StandardTableHorror.stories.tsx
@@ -1,3 +1,5 @@
+import { faAngleLeft } from "@fortawesome/free-solid-svg-icons/faAngleLeft";
+import { faAngleRight } from "@fortawesome/free-solid-svg-icons/faAngleRight";
 import { faCoffee } from "@fortawesome/free-solid-svg-icons/faCoffee";
 import { Box, Indent, Space, Text } from "@stenajs-webui/core";
 import { Chip, FlatButton, Icon, Tag } from "@stenajs-webui/elements";
@@ -598,6 +600,12 @@ const createSalesPerformanceStandardTableConfig = (
   columnGroups: {
     departures: {
       label: "Departures",
+      contentLeft: (
+        <FlatButton size={"small"} leftIcon={faAngleLeft} onClick={() => {}} />
+      ),
+      contentRight: (
+        <FlatButton size={"small"} leftIcon={faAngleRight} onClick={() => {}} />
+      ),
       columnOrder: [
         "link",
         "leg",